Removing "Total Cases Panel" and Method to HomePage
