+++ app/src/main/java/com/example/capstoneprojectk10/MainActivity.java
@@ -27,7 +27,6 @@ import timber.log.Timber;
 public class MainActivity extends AppCompatActivity {
 
     final Fragment homeFragment = new HomeFragment();
-    final Fragment statsFragment = new StatsFragment();
     final Fragment mapsFragment = new MapsFragment();
     final Fragment infoFragment = new InfoFragment();
     final FragmentManager fragmentManager = getSupportFragmentManager();
@@ -71,9 +70,6 @@ public class MainActivity extends AppCompatActivity {
                     switch (item.getItemId()) {
                         case R.id.nav_home:
                             fragment = new HomeFragment();
-                            break;
-                        case R.id.nav_stats:
-                            fragment = new StatsFragment();
                             break;
                         case R.id.nav_maps:
                             fragment = new MapsFragment();
