+++ app/src/main/res/values/colors.xml
@@ -9,7 +9,6 @@
     <color name="themeRed">#ff5959</color>
     <color name="themeGreen">#4cd97b</color>
     <color name="themeBlue">#4cb5ff</color>
-    <color name="themePurple">#9059ff</color>
     <color name="gray">#dddddd</color>
     <color name="gray700">#999999</color>
     <color name="splash">#1a1d3e</color>
